moved player code out of main and into player, also fixed vertical collision and jump physics. Still needs tweaks
